Refactoring NDT HTML5 driver unit tests (#22)
Refactoring mainly to get rid of custom-defined mock objects and replace them with uses
of the Mock library.



This should *not* introduce any significant changes in functionality. There are *some*

changes in functionality, including removing instances of autospec=True when there was

nothing to autospec and deleting mock behavior that was not relevant to the test.