Convert exceptions to errors, fix time recording
This makes a few changes that are a bit intermixed in the code:

1. Changes recording of start time and end time so that they record when the
underlying client process began and ended, not when the the actual tests
began (the revised behavior matches the spec).
2. Changes recording of s2c/c2s end_time values so that it actually waits for
the element it's waiting for before recording the time.
3. Changes two error cases so that instead of raising ValueError and killing
the process. One case was changed to a custom exception that kills the
process, as it should never occur. The other case was changed to a graceful
error that is recorded in results, but does not kill the process.
4. Removes the timeout handling logic from the result page parsing because
parsing the result page cannot time out (does not use WebDriverWait).

In addition, refactors html_driver a bit so that there is less mixing of
abstraction layers within functions.

Lastly, deletes the freezegun unit test as it was not providing additional
coverage over the final timing test and it's an additional unneeded
dependency.