Fixing time logging in HTML5 driver
This makes a few changes to recording timestamps during the test that are a
bit intermixed in the code:

1. Changes recording of start time and end time so that they record when the
underlying client process began and ended, not when the the actual tests began
(the revised behavior matches the spec).

2. Changes recording of s2c/c2s end_time values so that it actually waits for the
element it's waiting for before recording the time (it would previously record
the time without waiting, which was a bug and caused it to record incorrect
times).

In addition, refactors html_driver a bit so that there is less mixing of
abstraction layers within functions.

Lastly, deletes the freezegun unit test as it was not providing additional
coverage over the final timing test and it's an additional unneeded
dependency.