Adding a JSON decoder for NdtResult objects
Adds a JSON decoder so we can decode test results. Also adjusts the semantics
of the NdtResult & friends classes so they print better debug information
and support equality comparisons.

The client wrapper will not likely use this decoder directly, but it's
implemented here because consumers will need a way of decoding the JSON
output that the client wrapper produces and it seems easiest to have it
implemented once here rather than expect consumers to figure out how to
decode the JSON properly.