Adding CORS support to fake mlab-ns server
CORS support is needed if, for example, we have a web server on
localhost:8123 that wants to refer the client to a fake mlab-ns server on
localhost:8321. Without a CORS header, this is considered a violation of
same origin policy and the browser would block the response.

Adding CORS better simulates real mlab-ns.