Reducing impact of race condition in ReplayHTTPServer (#44)
* Reducing impact of race condition in ReplayHTTPServer

There is a race condition in ReplayHTTPServer in that one thread needs to
execute serve_forever, which blocks forever, but other threads need to know
that it has started serving.

I wasn't able to figure out a way to eliminate the race condition
completely, but I believe this reduces the impact to the point that it is
fixed for practical purposes. In practice, there will be a delay of several
milliseconds between the client making an HTTP request to the replay server
then making another HTTP request to the fake mlab-ns server.

This is mainly to fix flakiness in unit tests (issue #43), which it appears
to fix.