Adding a partial implementation of the Banjo driver (#47)
Adds a partial selenium driver for the Banjo client. Adds a unit test for
the currently implemented functionality.

Also fixes a minor issue with the TestResult constructor. It turns out that
using the default parameter of [] was causing all instances that used that
constructor to share the same list, which breaks unit tests. This fixes
the constructor and both client drivers' usage of the constructor.