Fixing client wrapper so it always quit()s Selenium (#55)
* Fixing client wrapper so it always quit()s Selenium

This fixes a bug in that the client_wrapper was just calling close() on
the Selenium driver, which closes the browser window, but does not free
resources associated with the browser driver. To do that, we need to call
quit().

This changes create_browser to a context manager so that callers can call
it in a with block that automatically calls quit() at the end. Updates the
unit tests to match.

Small refactoring of apply_patches_for_create_browser so that it's now shared
between both the banjo tests and the HTML5 tests.

Renames NdtClientTest to NdtClientTestCase