Fixing driver version check to work with Microsoft Edge (#57)
Microsoft's WebDriver for edge places the browser version information in the
'browserVersion' field instead of the 'version' field. This adds a function
that checks both fields if a 'version' field is not present and refactors
both the banjo and HTML5 clients to use this function instead of reading the
'version' field directly.