Fixing a crash in banjo driver
This fixes a crash in the banjo driver that occurred when load_url failed.
The failure would cause perform_test to return None, but callers expect it
to return a populated NdtResult object.

This fixes perform_test so that it always returns a populated NdtResult
even if load_url fails.