Fixing a bug in TestError timestamp calculation
This fixes a bug in the constructor for TestError. It turns out that the
default value of a function argument is an object, all callers share that
same object. This is undesirable for the TestError because each call should
result in a new timestamp of the current time.

This fixes the shared object and adds a unit test for the issue.