Fixing race condition in HTTPReplayServer (#59)
* Fixing race condition in HTTPReplayServer

The ReplayHTTPServer had an issue in that when it started up, it wouldn't
necessarily be serving HTTP requests yet because it started the actual
servers in background threads. This change explicitly waits until the
background threads are actually serving HTTP requests before returning.

Note: This is an uglier and more convoluted solution than I want, but it's
the best thing I can think of. I'm open to other solutions that eliminate
this race condition.