Adding fine-grained logging
This adds fine grained logging to the client wrapper. Specifically:

* Adds logging to the NDT HTML5 client wrapper
* Adds logging to the banjo driver
* Adds logging to fake mlab-ns
* Adds logging to top-level CLI module

A few notes:

We use the INFO level rather than the DEBUG level because once we use DEBUG,
Selenium's debug-level logging floods the logs. It might be possible to do
per-module debug-level, but for simplicity, it seemed easiest to start with
INFO.

There's a lot of redundancy in the logging in html5_driver.py. I didn't try
too hard to eliminate it because I expect that once we refactor html5_driver
to the same structure as banjo_driver, the logging will be as easy as it is
in banjo_driver.