Adding an HTTP replay generator
This continues the process of replacing the mitmdump dependency. This adds a
script that launches an HTTP proxy that captures traffic and writes it to a
YAML file so that it can be replayed locally.

There are deliberately no unit tests because they'd be fairly complex and this
feels like a utility script more than part of the core application, but we
can add unit tests if necessary.