Replaces mitmdump with a HttpResponse-based server (#73)
* Replaces mitmdump with a HttpResponse-based server

Changes the implementation of ReplayHTTPServer so that now it stores a
dictionary of HttpResponse objects to use for replaying traffic. This
eliminates the mitmdump dependency, so now everything can run within the same
process and we don't have to worry about orphaned mitmdump processes getting
left behind.

This also eliminates the need for FakeMLabNsServer because we can use a single
HTTP server to replay both the client responses and the mlab-ns responses.