Refactoring HTML5 driver
Refactoring the HTML5 driver to use a private helper class so that the
functions are not all passing around the same set of variables all the time.
This brings the HTML5 driver closer in style to the banjo driver.

This is a non-functional change (pure refactoring) so does not require any
changes to the unit tests.