Removing client_url parameter (#84)
This parameter is actually unnecessary. It was originally added for the HTML5
client, which needs a URL rather than a server, but the URL of the HTML5
client can be deduced, given the server FQDN.

Eliminating the client_url parameter simplifies semantics and allows similar
parameters to be used across different client types.